// frontend/src/app/core/models/api-route.model.ts
export enum ApiLabel {
  LOGIN = 'LOGIN',
  REFRESH = 'REFRESH',
  ABANDON_MEMBER_RESERVATION = 'ABANDON_MEMBER_RESERVATION',
  MODIFY_TIMETABLE = 'MODIFY_TIMETABLE',
  ACCEPT_THESIS_WITH_LECTURER = 'ACCEPT_THESIS_WITH_LECTURER',
  DELETE_THESIS_WITH_STUDENT = 'DELETE_THESIS_WITH_STUDENT',
  CORRECT_THESIS_WITH_LECTURER = 'CORRECT_THESIS_WITH_LECTURER',
  REJECT_THESIS_WITH_LECTURER = 'REJECT_THESIS_WITH_LECTURER',
  APPROVE_CHANGE_REQUEST = 'APPROVE_CHANGE_REQUEST',
  APPROVE_CLARIFICATION_REQUEST = 'APPROVE_CLARIFICATION_REQUEST',
  APPROVE_THESIS_WITH_COORDINATOR = 'APPROVE_THESIS_WITH_COORDINATOR',
  APPROVE_THESIS_WITH_COMMITTEE_MEMBER = 'APPROVE_THESIS_WITH_COMMITTEE_MEMBER',
  CONFIRM_MEMBER_RESERVATION = 'CONFIRM_MEMBER_RESERVATION',
  CONFIRM_PARTICIPATION_IN_RESERVATION = 'CONFIRM_PARTICIPATION_IN_RESERVATION',
  CREATE_CHANGE_REQUEST = 'CREATE_CHANGE_REQUEST',
  CREATE_CLARIFICATION_REQUEST = 'CREATE_CLARIFICATION_REQUEST',
  CREATE_RESERVATION = 'CREATE_RESERVATION',
  REJECT_RESERVATION = 'REJECT_RESERVATION',
  ACCEPT_RESERVATION = 'ACCEPT_RESERVATION',
  CREATE_THESIS = 'CREATE_THESIS',
  GET_CHANGE_REQUEST = 'GET_CHANGE_REQUEST',
  GET_CHANGE_REQUESTS = 'GET_CHANGE_REQUESTS',
  GET_CLARIFICATION_REQUEST = 'GET_CLARIFICATION_REQUEST',
  GET_CLARIFICATION_REQUESTS = 'GET_CLARIFICATION_REQUESTS',
  GET_DEPARTMENT = 'GET_DEPARTMENT',
  GET_DEPARTMENTS = 'GET_DEPARTMENTS',
  GET_DIPLOMA_SESSION = 'GET_DIPLOMA_SESSION',
  GET_DIPLOMA_SESSIONS = 'GET_DIPLOMA_SESSIONS',
  GET_EMPLOYEE = 'GET_EMPLOYEE',
  GET_EMPLOYEES = 'GET_EMPLOYEES',
  GET_FIELD_OF_STUDY = 'GET_FIELD_OF_STUDY',
  GET_FIELDS_OF_STUDY = 'GET_FIELDS_OF_STUDY',
  GET_RESERVATION = 'GET_RESERVATION',
  GET_RESERVATIONS = 'GET_RESERVATIONS',
  GET_STUDENT = 'GET_STUDENT',
  GET_STUDENTS = 'GET_STUDENTS',
  GET_THESES = 'GET_THESES',
  GET_THESIS = 'GET_THESIS',
  GET_TIMETABLE = 'GET_TIMETABLE',
  GET_TIMETABLES = 'GET_TIMETABLES',
  GET_USER = 'GET_USER',
  GET_NOTIFICATIONS = 'GET_NOTIFICATIONS',
  MODIFY_NOTIFICATION = 'MODIFY_NOTIFICATION',
  REJECT_THESIS_WITH_COORDINATOR = 'REJECT_THESIS_WITH_COORDINATOR',
  REJECT_THESIS_WITH_COMMITTEE_MEMBER = 'REJECT_THESIS_WITH_COMMITTEE_MEMBER',
  REJECT_CHANGE_REQUEST = 'REJECT_CHANGE_REQUEST',
  REJECT_CLARIFICATION_REQUEST = 'REJECT_CLARIFICATION_REQUEST',
  REQUEST_THESIS_CORRECTIONS = 'REQUEST_THESIS_CORRECTIONS',
}