// frontend/src/app/base/models/dto/topic-status.model.ts
export enum ThesisStatus {
  PROPOSED_BY_STUDENT = 'PROPOSED_BY_STUDENT',
  WAITING = 'WAITING',
  APPROVED_BY_COORDINATOR = 'APPROVED_BY_COORDINATOR',
  APPROVED_BY_COMMITTEE = 'APPROVED_BY_COMMITTEE',
  TO_CORRECT = 'TO_CORRECT',
  REJECTED_BY_LECTURER = 'REJECTED_BY_LECTURER',
  REJECTED_BY_COORDINATOR = 'REJECTED_BY_COORDINATOR',
  REJECTED_BY_COMMITTEE = 'REJECTED_BY_COMMITTEE',
}